Pixell

Documentation for Pixell.

Pixell.Enmap

Pixell.fullsky_geometry

Pixell.Enmap — Type

Map type, contains an AbstractArray and a WCS object, but behaves like the AbstractArray it contains for array operations. It only implements the subset of Base.Array operations which are common on maps. You should work with the data directly using enmap_instance.data if you need additional Array functions.

Pixell.fullsky_geometry — Method

fullsky_geometry([P=CarClenshawCurtis], res; shape = nothing, dims = ())

Arguments:

proj=CarClenshawCurtis: [optional] projection

res: resolution in radians. Passing a Number produces a square pixel. Passing a tuple with (ΔRA, ΔDEC) produces a rectangular pixel.

Keywords

shape::NTuple=nothing: shape of the map. If not specified, will be computed.

dims::NTuple=(): additional dimensions to append to the shape, such as (3,) for IQU to generate a map with (nx, ny, 3).

Returns:

shape::Tuple, wcs::WCSTransform: a tuple containing the shape of the map and the WCS

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(30/60))  # 30 arcmin pixel
((720, 361), WCSTransform(naxis=2,cdelt=[-0.5, 0.5],crval=[0.25, 0.0],crpix=[360.5, 181.0]))