Pixell

Documentation for Pixell.

Pixell.Enmap

Pixell.fullsky_geometry

Pixell.pix2sky

Pixell.pix2sky

Pixell.pix2sky!

Pixell.sky2pix

Pixell.sky2pix

Pixell.sky2pix!

Pixell.Enmap — Type

Map type, contains an AbstractArray and a WCS object, but behaves like the AbstractArray it contains for array operations. It only implements the subset of Base.Array operations which are common on maps. You should work with the data directly using enmap_instance.data if you need additional Array functions.

Pixell.fullsky_geometry — Method

fullsky_geometry([P=CarClenshawCurtis], res; shape = nothing, dims = ())

Generates a full-sky geometry.

Arguments:

proj=CarClenshawCurtis: [optional] projection

res: resolution in radians. Passing a Number produces a square pixel. Passing a tuple with (ΔRA, ΔDEC) produces a rectangular pixel.

Keywords

shape::NTuple=nothing: shape of the map. If not specified, will be computed.

dims::NTuple=(): additional dimensions to append to the shape, such as (3,) for IQU to generate a map with (nx, ny, 3).

Returns:

shape::Tuple, wcs::WCSTransform: a tuple containing the shape of the map and the WCS

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(30/60))  # 30 arcmin pixel
((720, 361), WCSTransform(naxis=2,cdelt=[-0.5, 0.5],crval=[0.25, 0.0],crpix=[360.5, 181.0]))

Pixell.pix2sky — Function

pix2sky(m::Enmap, pixcoords)

Convert 1-indexed pixels to sky coordinates. The output sky coordinates are determined by WCS, but usually are in units of degrees.

Arguments:

m::Enmap: the map that provides a coordinate system

pixcoords: pixcoords should be a 2-d array where "pixcoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

Returns:

Array: same shape as pixcoords

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(1))
       m = Enmap(rand(shape...), wcs)
julia> pix2sky(m, [1.0, 1.0])
2-element Vector{Float64}:
 180.0
 -90.0

Pixell.pix2sky! — Method

pix2sky!(m::Enmap, pixcoords, skycoords)

Convert 1-indexed pixels to sky coordinates, in-place. The output sky coordinates are determined by WCS, but usually are in units of degrees.

Arguments:

m::Enmap: the map that provides a coordinate system

pixcoords: pixel coordinates should be a 2-d array where "pixcoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

skycoords: output array for sky coordinates, must be same same as pixcoords

Returns:

Array: skycoords

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(1))
       m = Enmap(rand(shape...), wcs)
       pixcoords =  100 .* rand(2,4096 * 2)
       skycoords =  similar(pixcoords)

julia> pix2sky!(m, pixcoords, skycoords)

Pixell.pix2sky — Method

pix2sky(m::Enmap{T,N,AA,CarClenshawCurtis}, ra_pixel, dec_pixel)

Compute the sky position of a single position on the sky.

Only implemented for CAR (Clenshaw-Curtis variant) projections, so the input map is of type Enmap{T,N,AA,CarClenshawCurtis}. This takes pixel indices for RA and DEC, and returns a tuple containing the corresponding RA and DEC.

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(1))
       m = Enmap(rand(shape...), wcs)
julia> pix2sky(m, 30.0, 80.0)
(151.0, -11.0)

Pixell.sky2pix — Function

sky2pix(m::Enmap, skycoords)

Convert sky coordinates to 1-indexed pixels. The input sky coordinates are determined by WCS, but usually are in units of degrees.

Arguments:

m::Enmap: the map to obtain the coordinates from

skycoords: skycoords should be a 2-d array where "skycoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

Returns:

Array: same shape as skycoords

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(1))
       m = Enmap(rand(shape...), wcs)
julia> sky2pix(m, [30.0, 50.0])
2-element Vector{Float64}:
 151.0
 141.0

Pixell.sky2pix! — Method

sky2pix!(m::Enmap, skycoords, pixcoords)

Convert sky coordinates to 1-indexed pixels, in-place. The input sky coordinates are determined by WCS, but usually are in units of degrees.

Arguments:

m::Enmap: the map that provides a coordinate system

skycoords: sky coordinates should be a 2-d array where "skycoords[:, i]" is the i-th set of coordinates, or a 1-d array representing a single set of coordinates.

pixcoords: output array for pixel coordinates, must be same same as pixcoords

Returns:

Array: pixcoords

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(1))
       m = Enmap(rand(shape...), wcs)
       skycoords =  similar(pixcoords)
       pixcoords =  100 .* rand(2,4096 * 2)
julia> sky2pix!(m, skycoords, pixcoords)

Pixell.sky2pix — Method

sky2pix(m::Enmap{T,N,AA,CarClenshawCurtis}, ra, dec)

Compute 1-indexed pixels into sky coordinates.

Only implemented for CAR (Clenshaw-Curtis variant) projections. Takes RA and DEC and returns a tuple containing the corresponding pixel indices. If vectors of RA and DEC are given, then vectors of pixel indices will be returned.

Examples

julia> shape, wcs = fullsky_geometry(deg2rad(1))
       m = Enmap(rand(shape...), wcs)
julia> sky2pix(m, 30.0, 80.0)
(151.0, 171.0)